+++ packages/atomic/src/components/facets/atomic-facet/atomic-facet.tsx
@@ -86,9 +86,7 @@ import {MapProp} from '../../../utils/props-utils';
   styleUrl: 'atomic-facet.pcss',
   shadow: true,
 })
-export class AtomicFacet
-  implements InitializableComponent, BaseFacet<Facet, FacetState>
-{
+export class AtomicFacet implements InitializableComponent, BaseFacet<Facet> {
   @InitializeBindings() public bindings!: Bindings;
   public facet!: Facet;
   private dependenciesManager?: FacetConditionsManager;
@@ -109,6 +107,7 @@ export class AtomicFacet
   @Prop({mutable: true, reflect: true}) public facetId?: string;
   /**
    * The non-localized label for the facet.
+   * Used in the `atomic-breadbox` component through the bindings store.
    */
   @Prop({reflect: true}) public label = 'no-label';
   /**
@@ -148,6 +147,7 @@ export class AtomicFacet
    * The maximum number of results to scan in the index to ensure that the facet lists all potential facet values.
    * Note: A high injectionDepth may negatively impact the facet request performance.
    * Minimum: `0`
+   * Default: `1000`
    */
   @Prop() public injectionDepth = 1000;
   // @Prop() public customSort?: string; TODO: KIT-753 Add customSort option for facet
@@ -444,6 +444,7 @@ export class AtomicFacet
       return (
         <FacetPlaceholder
           numberOfValues={this.numberOfValues}
+          isCollapsed={this.isCollapsed}
         ></FacetPlaceholder>
       );
     }