+++ packages/atomic/src/components/facets/atomic-numeric-facet/atomic-numeric-facet.tsx
@@ -83,7 +83,7 @@ interface NumericRangeWithLabel extends NumericRangeRequest {
   shadow: true,
 })
 export class AtomicNumericFacet
-  implements InitializableComponent, BaseFacet<NumericFacet, NumericFacetState>
+  implements InitializableComponent, BaseFacet<NumericFacet>
 {
   @InitializeBindings() public bindings!: Bindings;
   public facet?: NumericFacet;
@@ -96,7 +96,7 @@ export class AtomicNumericFacet
 
   @BindStateToController('facet')
   @State()
-  public facetState?: NumericFacetState;
+  public facetState!: NumericFacetState;
   @BindStateToController('filter')
   @State()
   public filterState?: NumericFilterState;
@@ -111,6 +111,7 @@ export class AtomicNumericFacet
   @Prop({mutable: true, reflect: true}) public facetId?: string;
   /**
    * The non-localized label for the facet.
+   * Used in the `atomic-breadbox` component through the bindings store.
    */
   @Prop({reflect: true}) public label = 'no-label';
   /**
@@ -156,6 +157,7 @@ export class AtomicNumericFacet
    * The maximum number of results to scan in the index to ensure that the facet lists all potential facet values.
    * Note: A high injectionDepth may negatively impact the facet request performance.
    * Minimum: `0`
+   * Default: `1000`
    */
   @Prop({reflect: true}) public injectionDepth = 1000;
 
@@ -283,7 +285,7 @@ export class AtomicNumericFacet
 
   private formatValue(value: number) {
     try {
-      return this.formatter(value, this.bindings.i18n.languages);
+      return this.formatter(value, this.bindings.i18n.languages as string[]);
     } catch (error) {
       this.bindings.engine.logger.error(
         `atomic-numeric-facet facet value "${value}" could not be formatted correctly.`,
@@ -479,6 +481,7 @@ export class AtomicNumericFacet
       return (
         <FacetPlaceholder
           numberOfValues={this.numberOfValues}
+          isCollapsed={this.isCollapsed}
         ></FacetPlaceholder>
       );
     }